requirements.txt
google-cloud-storage
google-cloud-dlp
pandas-gbq
pandas
downcast
scipy
seaborn
matplotlib
jupyterlab
tqdm
streamlit
bokeh==2.4.1
fastapi
pydantic
uvicorn[standard]